requirements.txt
Django==3.0.8
django-import-export==2.2.0  
